+++ src/info/packages.c
@@ -1,8 +1,7 @@
-#include "info.h"
-#include "../config/config.h"
-#include "../utils.h"
-
+#include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
+#include <unistd.h>
 
 // just here to stop vscode from complaining about DT_DIR
 #ifndef __USE_MISC
@@ -10,11 +9,20 @@
 #endif
 
 #include <dirent.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
 #include <sqlite3.h>
 #include <sys/wait.h>
+
+#include "info.h"
+#include "../config/config.h"
+#include "../optdeps/optdeps.h"
+#include "../utils/utils.h"
+
+/*
+ * I know that I could probably use specific libraries (like ALPM
+ * or some RPM stuff) and add them as optional dependencies, however:
+ *   - I'm too lazy to do that
+ *   - that would be way slower than just counting directories
+ */
 
 // get the number of installed packages
 int packages(char *dest) {
@@ -25,16 +33,17 @@ int packages(char *dest) {
     unsigned count = 0;
     bool done = false;
 
-    #ifndef __APPLE__   // package managers that won't run on macOS
+    #ifdef __linux__   // package managers that won't run on macOS
         FILE *fp;
+        char *prefix = getenv("PREFIX");
 
         path[0] = 0;
-        if(getenv("PREFIX"))
-            strncpy(path, getenv("PREFIX"), 255);
+        if(prefix)
+            strncpy(path, prefix, 255);
         strncat(path, "/var/lib/pacman/local", 256-strlen(path));
         if(_pkg_pacman && (dir = opendir(path)) != NULL) {
             while((entry = readdir(dir)) != NULL)
-                if(entry->d_type == DT_DIR && strcmp(entry->d_name, ".") != 0 && strcmp(entry->d_name, "..") != 0)
+                if(entry->d_type == DT_DIR && entry->d_name[0] != '.')
                     ++count;
             closedir(dir);
 
@@ -45,8 +54,8 @@ int packages(char *dest) {
         }
 
         path[0] = 0;
-        if(getenv("PREFIX"))
-            strncpy(path, getenv("PREFIX"), 255);
+        if(prefix)
+            strncpy(path, prefix, 255);
         strncat(path, "/var/lib/dpkg/status", 256-strlen(path));
         if(_pkg_dpkg && (fp = fopen(path, "r")) != NULL) {
             char line[512];
@@ -67,10 +76,9 @@ int packages(char *dest) {
             }
         }
 
-        // could also use rpm APIs directly
         path[0] = 0;
-        if(getenv("PREFIX"))
-            strncpy(path, getenv("PREFIX"), 255);
+        if(prefix)
+            strncpy(path, prefix, 255);
         strncat(path, "/var/lib/rpm/rpmdb.sqlite", 256-strlen(path));
         if(_pkg_rpm && access(path, F_OK) == 0) {
             sqlite3 *db;
@@ -101,13 +109,13 @@ int packages(char *dest) {
         }
 
         path[0] = 0;
-        if(getenv("PREFIX"))
-            strncpy(path, getenv("PREFIX"), 255);
+        if(prefix)
+            strncpy(path, prefix, 255);
         strncat(path, "/var/lib/flatpak/runtime", 256-strlen(path));
         if(_pkg_flatpak && (dir = opendir(path)) != NULL) {
             count = 0;
             while((entry = readdir(dir)) != NULL)
-                if(entry->d_type == DT_DIR && strcmp(entry->d_name, ".") != 0 && strcmp(entry->d_name, "..") != 0)
+                if(entry->d_type == DT_DIR && entry->d_name[0] != '.')
                     ++count;
             closedir(dir);
 
@@ -119,13 +127,13 @@ int packages(char *dest) {
         }
 
         path[0] = 0;
-        if(getenv("PREFIX"))
-            strncpy(path, getenv("PREFIX"), 255);
+        if(prefix)
+            strncpy(path, prefix, 255);
         strncat(path, "/var/snap", 256-strlen(path));
         if(_pkg_snap && (dir = opendir(path)) != NULL) {
             count = 0;
             while((entry = readdir(dir)) != NULL)
-                if(entry->d_type == DT_DIR && strcmp(entry->d_name, ".") != 0 && strcmp(entry->d_name, "..") != 0)
+                if(entry->d_type == DT_DIR && entry->d_name[0] != '.')
                     ++count;
             closedir(dir);
 
@@ -137,7 +145,7 @@ int packages(char *dest) {
         }
     #endif
 
-    if(_pkg_brew && (access("/usr/local/bin/brew", F_OK) == 0 || access("/opt/homebrew/bin/brew", F_OK) == 0 || access("/bin/brew", F_OK) == 0)) {
+    if(_pkg_brew && (binary_in_path("brew"))) {
         char *args[] = {"brew", "--cellar", NULL};
         exec_cmd(str, 16, args);
 
@@ -159,12 +167,7 @@ int packages(char *dest) {
         }
     }
 
-
-    path[0] = 0;
-    if(getenv("PREFIX"))
-        strncpy(path, getenv("PREFIX"), 255);
-    strncat(path, "/bin/pip", 256-strlen(path));
-    if(_pkg_pip && access(path, F_OK) == 0) {
+    if(_pkg_pip && binary_in_path("pip")) {
         char *args[] = {"sh", "-c", "pip list 2>/dev/null | wc -l", NULL};
         exec_cmd(str, 16, args);