+++ src/utils/queue.c
@@ -1,8 +1,8 @@
-#include "queue.h"
-
 #include <stdbool.h>
 #include <stdio.h>
 #include <string.h>
+
+#include "queue.h"
 
 Queue *queue_with_size(size_t size) {
     Queue *q = malloc(sizeof(Queue));